macro: Move `test_casing` macro to separate module
